fix: move the unused base client to its own file, expose GoogleDriveClient and S3Client
